+++ src/light.rs
@@ -3,16 +3,16 @@ use animation::SetPosition;
 
 #[derive(Clone, Copy)]
 pub struct Light {
-    pub pos: Vec3<f64>,
-    pub color: Vec3<f64>,
+    pub pos: Vec3<f32>,
+    pub color: Vec3<f32>,
 }
 
 impl SetPosition for Light {
-    fn set_position(&mut self, pos: Vec3<f64>) {
+    fn set_position(&mut self, pos: Vec3<f32>) {
         self.pos = pos;
     }
 
-    fn get_position(&self) -> Vec3<f64> {
+    fn get_position(&self) -> Vec3<f32> {
         self.pos
     }
 }
