+++ src/material.rs
@@ -4,19 +4,19 @@ use light::Light;
 
 #[derive(Clone, Copy)]
 pub struct Material {
-    pub ambient: Vec3<f64>,
-    pub diffuse: Vec3<f64>,
-    pub specular: Vec3<f64>,
-    pub shininess: f64,
-    pub reflection: f64,
-    pub refraction: f64,
+    pub ambient: Vec3<f32>,
+    pub diffuse: Vec3<f32>,
+    pub specular: Vec3<f32>,
+    pub shininess: f32,
+    pub reflection: f32,
+    pub refraction: f32,
 }
 
-fn max(a: f64, b: f64) -> f64 {
+fn max(a: f32, b: f32) -> f32 {
     if a > b { a } else { b }
 }
 
-fn clamp(f: Vec3<f64>) -> Vec3<f64> {
+fn clamp(f: Vec3<f32>) -> Vec3<f32> {
     let mut h = f;
     h.x = if f.x  > 1.0 { 1.0 } else { f.x };
     h.y = if f.y  > 1.0 { 1.0 } else { f.y };
@@ -25,11 +25,11 @@ fn clamp(f: Vec3<f64>) -> Vec3<f64> {
 }
 
 impl Material {
-    pub fn compute_color(&self, ray: &Ray, tnear: f64, nhit: Vec3<f64>, light: &Light,
-                         reflected_color: Vec3<f64>, refracted_color: Vec3<f64>,
+    pub fn compute_color(&self, ray: &Ray, tnear: f32, nhit: Vec3<f32>, light: &Light,
+                         reflected_color: Vec3<f32>, refracted_color: Vec3<f32>,
                          light_shaded: bool)
-                         -> Vec3<f64> {
-        let mul = |l: &Vec3<f64>, r: &Vec3<f64>| -> Vec3<f64> {
+                         -> Vec3<f32> {
+        let mul = |l: &Vec3<f32>, r: &Vec3<f32>| -> Vec3<f32> {
             Vec3::new(l.x * r.x, l.y * r.y, l.z * r.z)
         };
