+++ src/surface.rs
@@ -58,7 +58,7 @@ impl Surface {
 }
 
 impl<'a> Division<'a> {
-    pub fn set_color(&self, x: u32, y: u32, color: Vec3<f64>) {
+    pub fn set_color(&self, x: u32, y: u32, color: Vec3<f32>) {
         let offset = ((y * self.surf.w + x) * 3) as usize;
         let pixels = &self.surf.pixels as *const Vec<u8> as *mut Vec<u8>;
         unsafe {
